[Bronze IV] Title: Большой удой, Time: 0 ms, Memory: 1112 KB -BaekjoonHub

+++ 백준/Bronze/26057. Большой удой/Большой удой.cc
@@ -1,7 +1,7 @@
 #include <cstdio>
 
 int main() {
-  long long L, T;
-  scanf("%lld%lld", &L, &T);
-  printf("%lld", 2 * T - L);
+  int L, T;
+  scanf("%d%d", &L, &T);
+  printf("%d", 2 * T - L);
 }
